Tweaked the field types showed on listings and used for search queries
In search queries we can use any type of array (array, json_array or simple_array)
because Doctrine just serialize/json_encode/implode the contents of the array and
thus, their contents are "searchable".

In listings we could of course display those types of fields thanks to the
|join Twig filter, but we exclude them by defualt because their contents
are usually very large and they easily break listing pages.