Reorganized the way the Request object is handled.
As said before, edit/new/show/... methods are not real Symfony controllers.
At first they were, and the Request argument is a leftover of that decision.
So let's remove this argument and use the Request provided by the real
indexAction() controller.

Someone may say that this change totally breaks compatibility and doesn't meet
SemVer philosophy. And all this is true. But the project was released yesterday
and nobody has done anything important with it, so nothing is going to break.
In my opinion it's OK to do this changes in the first hours of some project.
I won't make them next week or next month of course.